Revert "Complete Milemarker 2"
This reverts commit 1d7c4103d0d9a28c55d0ca5d22301a8522ef442d.
