+++ ep-support-portal/src/app/app.module.ts
@@ -3,17 +3,21 @@ import { NgModule } from '@angular/core';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { FormsModule } from '@angular/forms'; // NgModule Live here
 import { DataTablesModule } from 'angular-datatables';
+import { RouterModule, Routes } from '@angular/router';
 
 import { AppComponent } from './app.component';
-import { AppRoutingModule } from './app-routing.module';
-
 import { UsersTableComponent } from '././tables/users-table/users-table.component';
 import { CompanyComponent } from './pages/company/company.component';
 import { ReportsTableComponent } from './tables/reports-table/reports-table.component';
 import { PageNotFoundComponent } from './not-found.component';
 
 
-
+const appRoutes: Routes = [
+  { path: 'users', component: UsersTableComponent },
+  { path: 'reports', component: ReportsTableComponent },
+  { path: '', redirectTo: '/users', pathMatch: 'full' },
+  { path: '**', component: PageNotFoundComponent },
+]
 
 
 @NgModule({
@@ -29,7 +33,7 @@ import { PageNotFoundComponent } from './not-found.component';
     BrowserModule,
     FormsModule, // import the Forms Module before binding with [(NgModule)]
     DataTablesModule,
-    AppRoutingModule
+    RouterModule.forRoot(appRoutes)
   ],
   providers: [],
   bootstrap: [AppComponent]
